Added authentication to Category and UserGroup

+++ OutDistancing/src/main/java/com/skilldistillery/outdistancing/services/CategoryServiceImpl.java
@@ -8,13 +8,18 @@ import org.springframework.stereotype.Service;
 
 import com.skilldistillery.outdistancing.entities.Category;
 import com.skilldistillery.outdistancing.entities.Resource;
+import com.skilldistillery.outdistancing.entities.User;
 import com.skilldistillery.outdistancing.repositories.CategoryRepository;
+import com.skilldistillery.outdistancing.repositories.UserRepository;
 
 @Service
 public class CategoryServiceImpl implements CategoryService {
 
 	@Autowired
 	private CategoryRepository categoryRepo;
+	
+	@Autowired
+	private UserRepository userRepo;
 	
 	@Override
 	public List<Category> listAllCategories() {
@@ -34,18 +39,25 @@ public class CategoryServiceImpl implements CategoryService {
 	}
 
 	@Override
-	public Category createCategory(Category category) {
-		if (category != null) {
-			return categoryRepo.saveAndFlush(category);			
+	public Category createCategory(Category category, String username) {
+		User user = userRepo.findByUsername(username);
+		if (category != null && user != null) {
+			try {
+				return categoryRepo.saveAndFlush(category);	
+			} catch (Exception e) {
+				e.printStackTrace();
+				return null;
+			}		
 		}else {
 			return null;
 		}
 	}
 
 	@Override
-	public Category updateCategory(int categoryId, Category category) {
-        Optional<Category> optCategory = categoryRepo.findById(categoryId);
-        if (optCategory.isPresent()) {
+	public Category updateCategory(int categoryId, Category category, String username) {
+		User user = userRepo.findByUsername(username);
+		Optional<Category> optCategory = categoryRepo.findById(categoryId);
+        if (optCategory.isPresent() && user != null) {
             Category updateCategory = optCategory.get();
             updateCategory.setDescription(category.getDescription());
             updateCategory.setEnabled(category.isEnabled());
@@ -58,9 +70,10 @@ public class CategoryServiceImpl implements CategoryService {
 	}
 
 	@Override
-	public Boolean deleteCategory(int categoryId) {
+	public Boolean deleteCategory(int categoryId, String username) {
+		User user = userRepo.findByUsername(username);
 		Optional<Category> optCategory = categoryRepo.findById(categoryId);
-		if (optCategory.isPresent()) {
+		if (optCategory.isPresent() && user != null) {
 			Category updateCategory = optCategory.get();
 			if (updateCategory != null) {
 				categoryRepo.deleteById(categoryId);
@@ -72,9 +85,10 @@ public class CategoryServiceImpl implements CategoryService {
 	
 	
 	@Override
-	public Boolean changeCategoryEnabled(int categoryId) {
+	public Boolean changeCategoryEnabled(int categoryId, String username) {
+		User user = userRepo.findByUsername(username);
 		Optional<Category> optCategory = categoryRepo.findById(categoryId);
-        if (optCategory.isPresent()) {
+        if (optCategory.isPresent() && user != null) {
             Category updateCategory = optCategory.get();
             updateCategory.setEnabled(!updateCategory.isEnabled());
 			categoryRepo.save(updateCategory);
